Migrate spy_auto_quant to _base.auto_quant
